issue/#92-migrate-from-postgres-to-sqlite * update build.gradle dependencies * update application.yaml files * update compose files * update schema  * update repository methods to be sqlite dialect specific  * update unit tests

+++ backend/src/main/resources/databaseInit/schema.sql
@@ -2,6 +2,6 @@ DROP TABLE IF EXISTS finnish_words;
 
 CREATE TABLE IF NOT EXISTS finnish_words
 (
-    id      BIGSERIAL PRIMARY KEY,
+    id      INTEGER PRIMARY KEY,
     word    VARCHAR NOT NULL UNIQUE CHECK (LENGTH(word) >= 5 AND LENGTH(word) <= 7)
 );
